Polyphase filt, channal FILT_DC_AGC, ASK
Reset signals were entered into the files of the DS and AGS devices and a polyphase decimator filter was designed with variable decimation coefficients and the order of subfilters on which filters with distributed arithmetic. functions have been developed for calculating the polyphase matrix, bringing it to tables of distributed arithmetic for calculation in a shifted two's complement code. a channel for receiving one of the components of a polyphase signal is assembled. filter coefficients calculated